fixed warning and slightly better make file

+++ src/blur_filter.cpp
@@ -4,6 +4,7 @@
 
 #include <blur_filter.h>
 #include <bitmap.h>
+
 //tbb includes
 #include <tbb/parallel_for.h>
 
@@ -244,7 +245,7 @@ void blur_cuda(const uint8_t * h_source,
 {
 
 	//calculating the size of the arrya
-	int byte_size = width*height*3*sizeof(uint8_t);
+	int byte_size = width*height*3*(int)sizeof(uint8_t);
 
 	//declaring gpu pointers
 	uint8_t * d_source;
